@@ -4,7 +4,7 @@ import sys
 
 from setuptools import setup
 
-version = "1.0.0"
+version = "2.0.0"
 
 long_description = open('README.md').read()
 
@@ -25,11 +25,12 @@ setup(
     long_description_content_type="text/markdown",
     keywords='gdacs disasters earthquakes tropical-cyclones earthquakes floods',
     packages=['gdacs'],
-    python_requires=">=3.5",    
+    python_requires=">=3.6",    
     install_requires=[
         "requests>=2.10.0",
         "xmltodict",
         "cachetools",
+        "pydantic",
     ],
     classifiers=[
         "Development Status :: 3 - Alpha",
@@ -37,9 +38,10 @@ setup(
         'Operating System :: OS Independent',
         'License :: OSI Approved :: MIT License',
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
     ]
 )
