@@ -1,23 +1,17 @@
 import json
 import requests
-import xmltodict
 from os.path import join
 from cachetools import cached, TTLCache
 
-from gdacs.utils import GDACSAPIError
-from gdacs.utils import handle_xml, handle_geojson
-from gdacs.utils import download_shp
+from gdacs.schemas import GeoJSON
+from gdacs.utils import *
 
 
-CACHE_TTL = 300  # 5minutes
+CACHE_TTL = 300  # secs
 EVENT_TYPES = [None, 'TC', 'EQ', 'FL', 'VO', 'DR', 'WF']
 DATA_FORMATS = [None, 'xml', 'geojson', 'shp']
+LATEST_EVENTS_URL = 'https://www.gdacs.org/gdacsapi/api/events/geteventlist/EVENTS4APP'
 BASE_URL = "https://www.gdacs.org/datareport/resources"
-RSS_FEED_URLS = {
-    "default": "https://www.gdacs.org/xml/rss.xml",
-    "24h": "https://www.gdacs.org/xml/rss_24h.xml",
-    "7d": "https://www.gdacs.org/xml/rss_7d.xml"
-}
 
 
 class GDACSAPIReader:
@@ -30,28 +24,22 @@ class GDACSAPIReader:
     @cached(cache=TTLCache(maxsize=500, ttl=CACHE_TTL))
     def latest_events(self,
                       event_type: str = None,
-                      historical: str = 'default',
                       limit: int = None
                       ):
         """ Get latest events from GDACS RSS feed. """
         if event_type not in EVENT_TYPES:
             raise GDACSAPIError("API Error: Used an invalid `event_type` parameter in request.")
 
-        if historical not in RSS_FEED_URLS.keys():
-            raise GDACSAPIError("API Error: Used an invalid `historical` parameter in request.")
-
-        res = requests.get(RSS_FEED_URLS[historical])
+        res = requests.get(LATEST_EVENTS_URL)
         if res.status_code != 200:
             raise GDACSAPIError("API Error: GDACS RSS feed can not be reached.")
 
-        xml_parser = xmltodict.parse(res.content)
         events = [
-            item
-            for item in xml_parser["rss"]["channel"]["item"]
-            if event_type in [None, item["gdacs:eventtype"]]
+            event for event in res.json()['features']
+            if event_type in [None, event['properties']['eventtype']]
         ]
-
-        return json.loads(json.dumps(events[:limit]))
+        features = json.loads(json.dumps(events[:limit]))
+        return GeoJSON(features=features)
 
     @cached(cache=TTLCache(maxsize=500, ttl=CACHE_TTL))
     def get_event(self,
@@ -69,22 +57,29 @@ class GDACSAPIReader:
             raise GDACSAPIError("API Error: Used an invalid `data_format` parameter in request.")
 
         if source_format == 'geojson':
-            file_name = "geojson_{}_{}.geojson".format(event_id, episode_id)
-            geojson_path = join(BASE_URL, event_type, event_id, file_name).replace("\\", "/")
-            return handle_geojson(geojson_path)
-
+            return self.__get_geojson_event(event_type, event_id, episode_id)
         elif source_format == 'shp':
-            file_name = "Shape_{}_{}.zip".format(event_id, episode_id)
-            shp_path = join(BASE_URL, event_type, event_id, file_name).replace("\\", "/")
-            return download_shp(shp_path)
-
+            return self.__get_shp_event(event_type, event_id, episode_id)
         else:
-            if cap_file:
-                file_name = "cap_{}.xml".format(event_id)
-            elif not episode_id:
-                file_name = "rss_{}.xml".format(event_id)
-            else:
-                file_name = "rss_{}_{}.xml".format(event_id, episode_id)
+            return self.__get_xml_event(event_type, event_id, episode_id, cap_file)
 
-            xml_path = join(BASE_URL, event_type, event_id, file_name).replace("\\", "/")
-            return handle_xml(xml_path)
+    def __get_geojson_event(self, event_type: str, event_id: str, episode_id: str = None):
+        file_name = f"geojson_{event_id}_{episode_id}.geojson"
+        geojson_path = join(BASE_URL, event_type, event_id, file_name).replace("\\", "/")
+        return handle_geojson(geojson_path)        
+
+    def __get_shp_event(self, event_type: str, event_id: str, episode_id: str = None):
+        file_name = f"Shape_{event_id}_{episode_id}.zip"
+        shp_path = join(BASE_URL, event_type, event_id, file_name).replace("\\", "/")
+        return download_shp(shp_path)
+
+    def __get_xml_event(self, event_type: str, event_id: str, episode_id: str = None, cap_file: bool = False):
+        if cap_file:
+            file_name = f"cap_{event_id}.xml"
+        elif not episode_id:
+            file_name = f"rss_{event_id}.xml"
+        else:
+            file_name = f"rss_{event_id}_{episode_id}.xml"
+
+        xml_path = join(BASE_URL, event_type, event_id, file_name).replace("\\", "/")
+        return handle_xml(xml_path)
